test(degree): move delete test to integration test repo
