+++ libs/integration-tests/jest.config.ts
@@ -1,0 +1,15 @@
+/* eslint-disable */
+export default {
+  displayName: 'integration-tests',
+  preset: '../../jest.preset.js',
+  globals: {
+    'ts-jest': {
+      tsconfig: '<rootDir>/tsconfig.json',
+    },
+  },
+  transform: {
+    '^.+\\.[tj]s$': 'ts-jest',
+  },
+  moduleFileExtensions: ['ts', 'js', 'html'],
+  coverageDirectory: '../../coverage/libs/integration-tests',
+}
