+++ libs/integration-tests/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@modtree/integration-tests",
+  "version": "0.0.1",
+  "type": "commonjs",
+  "dependencies": {
+    "@modtree/types": "*"
+  }
+}
